+++ board-server/src/main/java/net/stzups/board/server/Server.java
@@ -23,7 +23,7 @@ import java.security.KeyStore;
  * Uses netty to create an HTTP/WebSocket server on the specified port
  */
 public class Server {
-    private static final int HTTP_PORT = 80;
+    private static final int HTTP_PORT = 8080;
     private static final int HTTPS_PORT = 443;
 
     private EventLoopGroup bossGroup;