+++ frontend/html/scripts/main.js
@@ -2,5 +2,5 @@ const Board = {};
 
 let script = document.createElement('script');
 script.type = "module"
-script.src = "/scripts/Board.js";
+script.src = "./scripts/Board.js";
 document.head.appendChild(script);