+++ CHANGELOG.md
@@ -75,3 +75,4 @@ SQL data persistence
 - Add environment variable config method
 - Fix send rate of client
 - Remove flatfile, add runtime (in memory only) and PostgresSQL database
+- Move frontend to independent HTTP server (nginx)